lambda/requirements.txt
# Strands Agents SDK - AWS open-source agent framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AWS SDK (boto3 is included in Lambda runtime, but pinning for local dev)
boto3>=1.34.0
botocore>=1.34.0

# HTTP requests for Lark API
requests>=2.31.0

# Type hints (for Python 3.12 compatibility)
typing-extensions>=4.9.0
